# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Versioned commerce API for the storefront and operations console
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/admin/summary:
    get:
      operationId: getAdminSummary
      tags: [admin]
      summary: Get operations summary
      security:
        - clerkSession: []
      responses:
        "200":
          description: Operations summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/admin/products:
    get:
      operationId: listAdminProducts
      tags: [admin]
      summary: List products for management
      security:
        - clerkSession: []
      responses:
        "200":
          description: Product catalog
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [admin]
      summary: Create a product
      security:
        - clerkSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /v1/admin/products/{productId}:
    patch:
      operationId: updateProduct
      tags: [admin]
      summary: Update a product
      security:
        - clerkSession: []
      parameters:
        - $ref: "#/components/parameters/ProductId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: archiveProduct
      tags: [admin]
      summary: Archive a product
      security:
        - clerkSession: []
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "204":
          description: Product archived
  /v1/admin/discounts:
    get:
      operationId: listDiscounts
      tags: [admin]
      summary: List dynamic discount codes
      security:
        - clerkSession: []
      responses:
        "200":
          description: Discount codes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Discount"
    post:
      operationId: createDiscount
      tags: [admin]
      summary: Create a dynamic discount code
      security:
        - clerkSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscountInput"
      responses:
        "201":
          description: Created discount
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Discount"
  /v1/admin/discounts/{discountId}:
    patch:
      operationId: updateDiscount
      tags: [admin]
      summary: Update a dynamic discount code
      security:
        - clerkSession: []
      parameters:
        - $ref: "#/components/parameters/DiscountId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscountUpdate"
      responses:
        "200":
          description: Updated discount
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Discount"
  /v1/admin/registrations:
    get:
      operationId: listRegistrationPolicies
      tags: [admin]
      summary: List first-time registration policies
      security:
        - clerkSession: []
      responses:
        "200":
          description: Registration policies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RegistrationPolicy"
    patch:
      operationId: updateRegistrationPolicy
      tags: [admin]
      summary: Update first-time registration policy
      security:
        - clerkSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationPolicyUpdate"
      responses:
        "200":
          description: Updated policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationPolicy"

  # ─── Checkout & Delivery (customer-facing) ───────────────────────────────
...
  /v1/admin/orders:
    get:
      operationId: listAdminOrders
      tags: [admin]
      summary: List all orders (operations console)
      security:
        - clerkSession: []
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: duration, in: query, schema: { type: string } }
        - { name: startDate, in: query, schema: { type: string } }
        - { name: endDate, in: query, schema: { type: string } }
        - { name: search, in: query, schema: { type: string } }
      responses:
        "200":
          description: Order ledger
          content:
            application/json:
              schema:
                type: array
                items: { type: object, additionalProperties: true }
  /v1/admin/orders/stats:
    get:
      operationId: getAdminOrderStats
      tags: [admin]
      summary: Aggregate order stats (operations console)
      security:
        - clerkSession: []
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalOrders: { type: number }
                  successfulOrders: { type: number }
                  pendingOrders: { type: number }
                  cancelledOrders: { type: number }
                  totalRevenueCents: { type: number }
  /v1/admin/orders/{id}/cancel:
    post:
      operationId: cancelAdminOrder
      tags: [admin]
      summary: Cancel any pending order (operations console)
      security:
        - clerkSession: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
  /v1/admin/shop-settings:
    get:
      operationId: getShopSettings
      tags: [admin]
      summary: Read shop settings
      security:
        - clerkSession: []
      responses:
        "200":
          description: Shop settings
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
    put:
      operationId: updateShopSettings
      tags: [admin]
      summary: Update shop settings
      security:
        - clerkSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { type: object, additionalProperties: true }

  # ─── Staff RBAC (Clerk-protected; login is open) ─────────────────────────
...
components:
  securitySchemes:
    clerkSession:
      type: http
      scheme: bearer
      bearerFormat: Clerk
    customerSession:
      type: http
      scheme: bearer
      bearerFormat: Opaque
    staffSession:
      type: http
      scheme: bearer
      bearerFormat: Opaque
  parameters:
    ProductId:
      name: productId
      in: path
      required: true
      schema:
        type: string
    DiscountId:
      name: discountId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Product:
      type: object
      required: [id, name, slug, description, priceCents, category, imageUrl, status, featured, inventory, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        priceCents: { type: number, minimum: 0 }
        compareAtPriceCents: { type: ["number", "null"], minimum: 0 }
        category: { type: string }
        imageUrl: { type: string }
        status: { type: string, enum: [active, draft, archived] }
        featured: { type: boolean }
        inventory: { type: number, minimum: 0 }
        createdAt: { type: string }
        updatedAt: { type: string }
    ProductInput:
      type: object
      required: [name, description, priceCents, category, imageUrl, inventory]
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        priceCents: { type: number, minimum: 0 }
        compareAtPriceCents: { type: ["number", "null"], minimum: 0 }
        category: { type: string }
        imageUrl: { type: string }
        status: { type: string, enum: [active, draft] }
        featured: { type: boolean }
        inventory: { type: number, minimum: 0 }
    ProductUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        priceCents: { type: number, minimum: 0 }
        compareAtPriceCents: { type: ["number", "null"], minimum: 0 }
        category: { type: string }
        imageUrl: { type: string }
        status: { type: string, enum: [active, draft, archived] }
        featured: { type: boolean }
        inventory: { type: number, minimum: 0 }
    Discount:
      type: object
      required: [id, code, type, value, minimumSubtotalCents, usageLimit, usageCount, startsAt, expiresAt, active, firstOrderOnly]
      properties:
        id: { type: string }
        code: { type: string }
        type: { type: string, enum: [percentage, fixed] }
        value: { type: number, minimum: 0 }
        minimumSubtotalCents: { type: number, minimum: 0 }
        usageLimit: { type: ["number", "null"], minimum: 0 }
        usageCount: { type: number, minimum: 0 }
        startsAt: { type: string }
        expiresAt: { type: ["string", "null"] }
        active: { type: boolean }
        firstOrderOnly: { type: boolean }
    DiscountInput:
      type: object
      required: [code, type, value, minimumSubtotalCents, startsAt, firstOrderOnly]
      properties:
        code: { type: string, minLength: 3 }
        type: { type: string, enum: [percentage, fixed] }
        value: { type: number, minimum: 0 }
        minimumSubtotalCents: { type: number, minimum: 0 }
        usageLimit: { type: ["number", "null"], minimum: 0 }
        startsAt: { type: string }
        expiresAt: { type: ["string", "null"] }
        active: { type: boolean }
        firstOrderOnly: { type: boolean }
    DiscountUpdate:
      type: object
      properties:
        value: { type: number, minimum: 0 }
        minimumSubtotalCents: { type: number, minimum: 0 }
        usageLimit: { type: ["number", "null"], minimum: 0 }
        startsAt: { type: string }
        expiresAt: { type: ["string", "null"] }
        active: { type: boolean }
        firstOrderOnly: { type: boolean }
...
    AdminSummary:
      type: object
      required: [activeProducts, draftProducts, liveDiscounts, firstOrderRegistrations, inventoryValueCents, recentActivity]
      properties:
        activeProducts: { type: number }
        draftProducts: { type: number }
        liveDiscounts: { type: number }
        firstOrderRegistrations: { type: number }
        inventoryValueCents: { type: number }
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
    Activity:
      type: object
      required: [id, label, detail, timestamp]
      properties:
        id: { type: string }
        label: { type: string }
        detail: { type: string }
        timestamp: { type: string }
    RegistrationPolicy:
      type: object
      required: [id, name, description, offerCode, active, windowDays, registrationsCount]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        offerCode: { type: string }
        active: { type: boolean }
        windowDays: { type: number, minimum: 0 }
        registrationsCount: { type: number, minimum: 0 }
    RegistrationPolicyUpdate:
      type: object
      properties:
        active: { type: boolean }
        offerCode: { type: string }
        windowDays: { type: number, minimum: 0 }
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }